hotelpilot/requirements.txt
# HotelPilot Requirements
# Multi-Agent Hotel Operations System

# Core Google ADK
google-adk>=1.15.1

# Google AI and Cloud
google-genai>=1.21.1
google-cloud-storage>=2.18.0
google-api-python-client>=2.157.0

# Async and Networking
aiohttp>=3.9.0
httpx>=0.28.1
websockets>=15.0.1

# Data Processing
pydantic>=2.0
python-dateutil>=2.9.0
PyYAML>=6.0.2

# Utilities
python-dotenv>=1.0.0
tenacity>=8.0.0
typing-extensions>=4.5

# Logging and Monitoring
opentelemetry-api>=1.31.0
opentelemetry-sdk>=1.31.0

# Development Tools (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0
mypy>=1.5.0

# Production Dependencies (optional)
# stripe>=7.0.0  # For payment processing
# twilio>=8.0.0  # For SMS/Voice
# redis>=5.0.0  # For caching
# sqlalchemy>=2.0  # For database
# alembic>=1.12.0  # For migrations